+++ packages/sdk/src/index.ts
@@ -23,5 +23,5 @@ export type {
   BrebClient as SwapBrebClient,
   ExternalUsBankSwapClient,
 } from '@bloque/sdk-swap';
-export type { BloqueClients } from './bloque';
+export type { BloqueClients, UpdateOriginMetadataOptions } from './bloque';
 export { SDK } from './bloque';